Modified features by Nikhil

+++ frontend/src/components/patient/DeletePatient.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { useEffect } from 'react';
-import { useParams, useNavigate } from 'react-router';
+import { useParams, useNavigate } from 'react-router-dom';
 
 export default function DeletePatient(){
     
@@ -8,21 +8,27 @@ export default function DeletePatient(){
     let navigate=useNavigate();
     useEffect(()=>{
         
+        
+        if (!window.confirm("Are you sure you want to delete this patient?")) {
+                navigate("/patient");
+                return;
+        }
+
         let url="http://localhost:9002/api/deletePatient/"+pid;
         axios.delete(url)
         .then((res)=>{
             alert("Deleted successfully");
-            navigate("/patient");
+            navigate("/patient/display");
 
         })
         .catch((err)=>{
-            console.error(err);
+            alert(err.message);
         })
-    })
+    },[pid,navigate])
 
     return(
         <div>
-            <h3>This is delete Patient</h3>
+            <h3>Deleting Patient...</h3>
         </div>
     )
 }
